Areas are now deleted from a world copy
The world is duplicated and the back-up copy is used to replace areas
that are deleted. This is not as efficient as it could be right now, but
it works. The main issue is that lighting updates do not occur due to
NMS fast pasting, so dark areas can occur on the landscape.