Added new admin command to place and record structures
This command will save the structure in the structures.yml file so that
it can be made when players make a new Box. Note that the server needs
to know about these before they can be loaded (API bug?) by issuing the
/place command or having had them load sometime previously in the
server, e.g., by going into a biome. Just needs to be done once.